debug new timeout

+++ src/server1.c
@@ -287,8 +287,8 @@ int send_file(int sockfd, struct sockaddr_in *client_addr, socklen_t client_addr
                             struct timespec end_time;
                             clock_gettime(CLOCK_REALTIME, &end_time);
                             int duree_emition = (int) ((begin_timeouts[loop_max].tv_sec*100000 + begin_timeouts[loop_max].tv_nsec/1000) - (begin_timeouts[i].tv_sec*100000 + begin_timeouts[i].tv_nsec/1000)); // durée d'emission des paquets en us
-                            int start_time_us = (int) ((begin_timeouts[i].tv_sec*100000 + begin_timeouts[i].tv_nsec/1000));
-                            int end_time_us = (int) ((end_time.tv_sec*100000 + end_time.tv_nsec/1000));
+                            long long int start_time_us = (long long int ) ((begin_timeouts[i].tv_sec*100000 + begin_timeouts[i].tv_nsec/1000));
+                            long long int  end_time_us = (long long int ) ((end_time.tv_sec*100000 + end_time.tv_nsec/1000));
                             RTT = end_time_us - start_time_us - duree_emition; 
                             printf("RTT : %d us\n", RTT);
                             printf("Start time for element %d : %d\n Stop time : %d \n", i, start_time_us, end_time_us);
